increase accelerometer field width to 16 bits

+++ ehm-common/messagetypes.h
@@ -42,9 +42,9 @@ typedef PACKEDSTRUCT {
 } GPSMessage;
 
 typedef PACKEDSTRUCT {
-	int8_t x;
-	int8_t y;
-	int8_t z;
+	int16_t x;
+	int16_t y;
+	int16_t z;
 } AccelerometerMessage;
 
 typedef PACKEDSTRUCT {
